move resetEnvironment accounts reset to vm provider
