+++ src/blockchain/blockchain.js
@@ -11,7 +11,7 @@ const Personal = require('web3-eth-personal')
 const Web3 = require('web3')
 
 const async = require('async')
-const { BN, privateToAddress, isValidPrivate, stripHexPrefix, toChecksumAddress } = require('ethereumjs-util')
+const { BN, privateToAddress, isValidPrivate, stripHexPrefix } = require('ethereumjs-util')
 const crypto = require('crypto')
 const { EventEmitter } = require('events')
 
@@ -350,15 +350,8 @@ class Blockchain {
   }
 
   resetEnvironment () {
-    this.providers.vm.accounts = {}
-    if (this.executionContext.isVM()) {
-      this._addAccount('3cd7232cd6f3fc66a57a6bedc1a8ed6c228fff0a327e169c2bcc5e869ed49511', '0x56BC75E2D63100000')
-      this._addAccount('2ac6c190b09897cd8987869cc7b918cfea07ee82038d492abce033c75c1b1d0c', '0x56BC75E2D63100000')
-      this._addAccount('dae9801649ba2d95a21e688b56f77905e5667c44ce868ec83f82e838712a2c7a', '0x56BC75E2D63100000')
-      this._addAccount('d74aa6d18aa79a05f3473dd030a97d3305737cbc8337d940344345c1f6b72eea', '0x56BC75E2D63100000')
-      this._addAccount('71975fbf7fe448e004ac7ae54cad0a383c3906055a65468714156a07385e96ce', '0x56BC75E2D63100000')
-    }
-    // TODO: most params here can be refactored away in txRunner
+    this.getCurrentProvider().resetEnvironment()
+   // TODO: most params here can be refactored away in txRunner
     this.txRunner = new TxRunner(this.providers.vm.accounts, {
       // TODO: only used to check value of doNotShowTransactionConfirmationAgain property
       config: this.config,
@@ -387,7 +380,7 @@ class Blockchain {
     if (this.executionContext.getProvider() !== 'vm') {
       throw new Error('plugin API does not allow creating a new account through web3 connection. Only vm mode is allowed')
     }
-    this._addAccount(privateKey, balance)
+    this.providers.vm._addAccount(privateKey, balance)
     const privKey = Buffer.from(privateKey, 'hex')
     return '0x' + privateToAddress(privKey).toString('hex')
   }
@@ -398,7 +391,7 @@ class Blockchain {
       do {
         privateKey = crypto.randomBytes(32)
       } while (!isValidPrivate(privateKey))
-      this._addAccount(privateKey, '0x56BC75E2D63100000')
+      this.providers.vm._addAccount(privateKey, '0x56BC75E2D63100000')
       return cb(null, '0x' + privateToAddress(privateKey).toString('hex'))
     }
     if (!this.config.get('settings/personal-mode')) {
@@ -407,30 +400,6 @@ class Blockchain {
     passwordPromptCb((passphrase) => {
       this.executionContext.web3().personal.newAccount(passphrase, cb)
     })
-  }
-
-  /** Add an account to the list of account (only for Javascript VM) */
-  _addAccount (privateKey, balance) {
-    if (!this.executionContext.isVM()) {
-      throw new Error('_addAccount() cannot be called in non-VM mode')
-    }
-
-    if (this.providers.vm.accounts) {
-      privateKey = Buffer.from(privateKey, 'hex')
-      const address = privateToAddress(privateKey)
-
-      // FIXME: we don't care about the callback, but we should still make this proper
-      let stateManager = this.executionContext.vm().stateManager
-      stateManager.getAccount(address, (error, account) => {
-        if (error) return console.log(error)
-        account.balance = balance || '0xf00000000000000001'
-        stateManager.putAccount(address, account, (error) => {
-          if (error) console.log(error)
-        })
-      })
-
-      this.providers.vm.accounts[toChecksumAddress('0x' + address.toString('hex'))] = { privateKey, nonce: 0 }
-    }
   }
 
   /** Get the balance of an address, and convert wei to ether */
